fix(ai): resolve golangci-lint issues in ai/ module
Fix 79+ golangci-lint issues across migrated ai/ directory:
- errcheck: Add _ = prefix and nolint comments for unchecked errors
- gocritic: Add explanations for nolint directives
- staticcheck: Fix error string capitalization
- gofmt: Fix formatting issues with nolint comments
- unconvert: Remove unnecessary int64 conversions
- prealloc: Pre-allocate slices with known capacity
- ineffassign: Remove unused variable assignments
- nilerr: Add nolint comments for intentional fallback behaviors

+++ ai/metrics/service.go
@@ -118,8 +118,8 @@ func (s *Service) GetStats(ctx context.Context, timeRange TimeRange) (*AgentMetr
 		memRequests := stats.RequestCount - totalDBRequests
 		if memRequests > 0 && stats.LatencyP50 > 0 {
 			// Weighted average of memory and DB P50/P95
-			memP50Weighted := int64(stats.LatencyP50.Milliseconds()) * memRequests
-			memP95Weighted := int64(stats.LatencyP95.Milliseconds()) * memRequests
+			memP50Weighted := stats.LatencyP50.Milliseconds() * memRequests
+			memP95Weighted := stats.LatencyP95.Milliseconds() * memRequests
 			stats.LatencyP50 = time.Duration((memP50Weighted+totalP50Weighted)/stats.RequestCount) * time.Millisecond
 			stats.LatencyP95 = time.Duration((memP95Weighted+totalP95Weighted)/stats.RequestCount) * time.Millisecond
 		} else {
